ZERO-185 [Rename]: 스토어 이름 및 경로 변경
