+++ src/components/team/member/ExileUserModal.tsx
@@ -1,5 +1,5 @@
 import { deleteMemberById } from '@/src/api/team/memberAPI';
-import { useTeamStore } from '@/src/stores/teamStore';
+import { useTeamStore } from '@/src/stores/useTeamStore';
 import Button from '@components/@shared/Button';
 import { Modal } from '@components/@shared/Modal';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
